safari-534.53-branch/LayoutTests/http/tests/security/isolatedWorld/didClearWindowObject-expected.txt

PASS: __worldID is undefined in normal world.
PASS: __worldID in isolated world should be 1 and is.
